+++ components/common/SearchBar.tsx
@@ -1,45 +1,11 @@
 'use client';
 
-import { usePathname, useRouter, useSearchParams } from 'next/navigation';
 import { useState } from 'react';
+
+import { useMyURLSearchParams } from '@/utils/search';
 
 interface SearchBarProps {
   margin?: string;
-}
-
-type QueryKey = 'keyword' | 'tag';
-
-export function useMyURLSearchParams() {
-  const searchParams = useSearchParams();
-  const pathname = usePathname();
-  const router = useRouter();
-
-  const setSearchParams = (key: QueryKey, value: string, replace: boolean = true) => {
-    const params = new URLSearchParams(Array.from(searchParams.entries()));
-
-    if (!value) {
-      params.delete(key);
-    } else if (replace) {
-      params.set(key, value);
-    } else {
-      params.append(key, value);
-    }
-
-    const query = params.toString();
-    const pathWithQuery = query ? `${pathname}?${query}` : pathname;
-    router.push(pathWithQuery);
-  };
-
-  const deleteSearchParams = (key: QueryKey, value?: string) => {
-    const params = new URLSearchParams(Array.from(searchParams.entries()));
-
-    if (!value) {
-      params.delete(key);
-    } else {
-    }
-  };
-
-  return { setSearchParams, deleteSearchParams } as const;
 }
 
 export default function SearchBar({ margin = '' }: SearchBarProps) {
